# artifacts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ONCE Internal OS — development increment 1",
    "version": "0.1.0-dev.1",
    "description": "Paths and strict request schemas are generated from runtime routes. Response shapes are currently TypeScript DTOs; this is not a complete response-schema validator."
  },
  "paths": {
...
    "/api/v1/auth/activate": {
      "post": {
        "operationId": "auth.activate",
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "token": {
                    "type": "string",
                    "minLength": 32,
                    "maxLength": 100
                  },
                  "password": {
                    "type": "string",
                    "minLength": 12,
                    "maxLength": 256
                  }
                },
                "required": [
                  "token",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Allowlisted response DTO; see src/dto.ts"
          },
          "default": {
            "description": "Sanitized error with code, message, requestId"
          }
        },
        "x-permission": null,
        "x-mode": "AUTH"
      }
    },
...
    "/api/v1/memberships/{id}/disable": {
      "post": {
        "operationId": "member.disable",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 128
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "expectedRevision": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 9007199254740991
                  }
                },
                "required": [
                  "expectedRevision"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Minimal command receipt; import commit acknowledges enqueue only"
          },
          "default": {
            "description": "Sanitized error with code, message, requestId"
          }
        },
        "x-permission": "members.manage",
        "x-mode": "COMMAND"
      }
    },
    "/api/v1/memberships/{id}/permissions": {
      "patch": {
        "operationId": "member.permissions",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 128
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "expectedRevision": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 9007199254740991
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "ADMIN",
                      "EDITOR",
                      "REVIEWER",
                      "VIEWER"
                    ]
                  },
                  "extraPermissions": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "sensitive.read",
                        "sensitive.write"
                      ]
                    },
                    "minItems": 0,
                    "maxItems": 2
                  }
                },
                "required": [
                  "expectedRevision",
                  "role",
                  "extraPermissions"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Minimal command receipt; import commit acknowledges enqueue only"
          },
          "default": {
            "description": "Sanitized error with code, message, requestId"
          }
        },
        "x-permission": "members.manage",
        "x-mode": "COMMAND"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "once_session"
      }
    }
  }
}